Revert "[FIX] 시설 상세정보 출력화면 구현 중"
