--- src/components/Reserve/common/CenteredContainer.jsx
@@ -1,13 +1,0 @@
-import { Container } from "@mui/material";
-
-const CenteredContainer = ({ children }) => {
-    return (
-        <Container
-            sx={{ height: "calc(100vh - 250px)", display: "flex", alignItems: "center", justifyContent: "center" }}
-        >
-            {children}
-        </Container>
-    );
-};
-
-export default CenteredContainer;
