+++ src/components/Reserve/content/ListContent.jsx
@@ -1,90 +1,115 @@
 import React, { useEffect, useState } from "react";
-import { Container, Typography, Button } from "@mui/material";
-import Loading from "../../Global/Loading.jsx";
-import useApiFetch from "../../../hook/Reserve/useApiFetch.js";
+import { Container, Typography } from "@mui/material";
+import FacilityCards from "./FacilityCards.jsx";
+import { useReserveContext } from "../../../context/ReserveContext.jsx";
+import useScrollRestore from "../../../hook/Reserve/useScrollRestore.js";
 import useInfiniteScroll from "../../../hook/Reserve/useInfiniteScroll.js";
-import CenteredContainer from "../common/CenteredContainer.jsx";
+import useTodayTimer from "../../../hook/Reserve/useTodayTimer.js";
+import useGeolocation from "../../../hook/Reserve/useGeolocation.js";
+import { getFacilityListsToReserve } from "../../../services/reserveService.js";
+import Loading from "../../Global/Loading.jsx";
 
-const ListContent = ({
-    apiFn,
-    params = {},
-    deps = [],
-    isList = true,
-    noDataComment = "결과가 없습니다.",
-    itemComponent: ItemComponent,
-}) => {
-    const [isLast, setIsLast] = useState(false);
-    const [legacy, setLegacy] = useState([]);
-    const [beforePage, setBeforePage] = useState(0);
+const ListContent = () => {
+    const {
+        category,
+        sortBy,
+        data,
+        setData,
+        page,
+        setPage,
+        size,
+        getScrollY,
+        setScrollY,
+        location,
+        setLocation,
+        noData,
+        setNoData,
+        last,
+        setLast,
+    } = useReserveContext();
 
-    useInfiniteScroll(isLast, params.setPage);
+    const [loading, setLoading] = useState(false); // 로딩 상태
+    const [error, setError] = useState(null); // 에러 상태
 
-    const { data, loading, error, last, noData } = useApiFetch({
-        apiFn,
-        params,
-        deps: [deps],
-    });
+    useScrollRestore(getScrollY, setScrollY);
+    useInfiniteScroll(setPage);
+    const today = useTodayTimer().locale("en").format("ddd");
+
+    const { latitude, longitude } = useGeolocation();
 
     useEffect(() => {
-        if (data && data.length > 0) {
-            if (isList) {
-                if (params.page - beforePage === 1) {
-                    setLegacy((prev) => [...prev, ...data]);
-                } else {
-                    setLegacy(data);
-                }
-                setBeforePage(params.page);
-            } else {
-                setLegacy(data);
-            }
+        if (latitude && longitude) {
+            setLocation({ latitude, longitude });
         }
-    }, [data]);
+    }, [latitude, longitude, setLocation]);
 
-    // last 상태 변경 감지
     useEffect(() => {
-        if (last) setIsLast(true);
-        console.log("ListContent - 마지막 페이지 감지");
-    }, [last]);
+        const fetchFacilities = async () => {
+            setLoading(true); // 로딩 시작
+            setError(null); // 에러 초기화
+            try {
+                if (noData || last) return;
+                const result = await getFacilityListsToReserve({
+                    category,
+                    sortBy,
+                    page,
+                    size,
+                    location, // location 객체 전달
+                    today, // 오늘 날짜 전달
+                });
 
-    console.log("ListContent 응답 데이터: " + data.length + " 개");
+                if (!result || !Array.isArray(result.content)) {
+                    setError("서버 데이터 형식 오류");
+                }
 
-    if (noData)
+                if (result.content.length === 0 && page === 0) {
+                    setNoData(true);
+                } else {
+                    setNoData(false);
+                }
+                const { content, last: isLast } = result;
+                setLast(isLast);
+
+                if (page === 0) {
+                    setData(content); // 첫 페이지는 데이터 초기화
+                } else {
+                    setData((prev) => [...prev, ...content]); // 이후 페이지는 누적
+                }
+            } catch (err) {
+                setError(err + "시설 목록을 불러오는 데 실패했습니다."); // 에러 처리
+            } finally {
+                setLoading(false); // 로딩 종료
+            }
+        };
+
+        if (location.latitude && location.longitude) {
+            fetchFacilities(); // 위치 정보가 있을 때만 데이터 가져오기
+        }
+    }, [page, sortBy, category, location]); // 의존성 배열에 필요한 값들 추가
+
+    if (loading) {
         return (
-            <CenteredContainer>
-                <Typography>{noDataComment}</Typography>
-            </CenteredContainer>
-        );
-
-    if (loading && (!data || data.length === 0)) {
-        return (
-            <CenteredContainer>
+            <Container
+                sx={{ height: "calc(100vh - 250px)", display: "flex", alignItems: "center", justifyContent: "center" }}
+            >
                 <Loading />
-            </CenteredContainer>
+            </Container>
         );
     }
-
     if (error) {
         return (
-            <CenteredContainer>
+            <Container
+                sx={{ height: "calc(100vh - 250px)", display: "flex", alignItems: "center", justifyContent: "center" }}
+            >
                 <Typography color="error">{error}</Typography>
-                <Button variant="contained" sx={{ mt: 2 }} onClick={() => params.setPage(0)}>
-                    다시 시도
-                </Button>
-            </CenteredContainer>
+            </Container>
         );
     }
-
     return (
         <Container disableGutters sx={{ paddingTop: "160px" }}>
-            {legacy && legacy.map((item) => <ItemComponent key={item.id} item={item} />)}
-
-            {loading && data && data.length > 0 && <Loading />}
-
-            {isLast && data && data.length > 0 && (
-                <Typography variant="body2" color="textSecondary" sx={{ my: 2 }}>
-                    마지막 항목입니다
-                </Typography>
-            )}
+            {data.map((item) => (
+                <FacilityCards key={item.id} item={item} />
+            ))}
         </Container>
     );
 };
