+++ src/components/Reserve/readme.md
@@ -1,4 +1,1 @@
-# Reserve 디렉터리 구조 설명
-## App.jsx 라우터 설정은 /reserve/*
-## 단일 페이지. 레이아웃 될 컴포넌트만 교체하는 방식
-## App.jsx(루트 라우터) -> Reserve.jsx(섹션 라우트) -> ~View.jsx(화면 컴포넌트) -> ~Content.jsx(화면 컴포넌트) -> filter.jsx, item.jsx 등(그 외 세부 컴포넌트)
+Reserve 컴포넌트가 들어가는 폴더
