+++ src/components/Reserve/result/ReservationSuccess.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const ReservationSuccess = () => {
+    return (
+        <Box>
+            <Box></Box>
+        </Box>
+    );
+};
+
+export default ReservationSuccess;
