+++ src/hook/Reserve/useInfiniteScroll.js
@@ -1,13 +1,12 @@
 import { useEffect } from "react";
 import { throttle } from "lodash";
 
-const useInfiniteScroll = (isLast, setPage) => {
+const useInfiniteScroll = (setPage) => {
     useEffect(() => {
-        if (setPage === undefined || setPage === null) return;
         // 쓰로틀링을 사용하여 스크롤 이벤트 최적화
         const handleInfiniteScroll = throttle(() => {
             const isBottom = window.innerHeight + window.scrollY >= document.body.offsetHeight - 10;
-            if (!isLast && isBottom) {
+            if (isBottom) {
                 setPage((prev) => prev + 1);
             }
         }, 200); // 200ms 간격으로 처리
@@ -17,7 +16,7 @@ const useInfiniteScroll = (isLast, setPage) => {
         return () => {
             window.removeEventListener("scroll", handleInfiniteScroll);
         };
-    }, [isLast, setPage]);
+    }, [setPage]);
 };
 
 export default useInfiniteScroll;
