+++ src/pages/Reserve/Reserve.jsx
@@ -1,87 +1,56 @@
 import React from "react";
+import CategoryFilter from "../../components/Reserve/filter/CategoryFilter.jsx";
+import SortFilter from "../../components/Reserve/filter/SortFilter";
+import ListContent from "../../components/Reserve/content/ListContent.jsx";
+import { Box, Button, Container, Divider } from "@mui/material";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
-import {
-    FacilityDetailProvider,
-    FacilityListProvider,
-    ReserveDetailProvider,
-    ReserveListProvider,
-    ReserveRouteProvider,
-    useReserveRouteContext,
-} from "../../context/ReserveContext.jsx";
-import FacilityDetailView from "../../components/Reserve/view/FacilityDetailView.jsx";
-import ReserveListView from "../../components/Reserve/view/ReserveListView.jsx";
-import ReserveDetailView from "../../components/Reserve/view/ReserveDetailView.jsx";
-import ReserveResultView from "../../components/Reserve/view/ReserveResultView.jsx";
-import ReviewInputFormView from "../../components/Reserve/view/ReviewInputFormView.jsx";
-import { Container, Typography } from "@mui/material";
-import TitleBar from "../../components/Global/TitleBar.jsx";
-import CenteredContainer from "../../components/Reserve/common/CenteredContainer.jsx";
-import { useParams } from "react-router-dom";
-import FacilityListView from "../../components/Reserve/view/FacilityListView.jsx";
+import { useNavigate } from "react-router-dom";
+import { ReserveProvider } from "../../context/ReserveContext";
 
-const Reserve = () => {
-    const RenderContent = () => {
-        const { path } = useReserveRouteContext();
-        const id = useParams();
+const ReserveContent = () => {
+    const navigate = useNavigate();
 
-        if (path.startsWith("/reserve/facility/") && id)
-            return (
-                <FacilityDetailProvider>
-                    <TitleBar name="편의시설 상세정보" />
-                    <FacilityDetailView id={id} />
-                </FacilityDetailProvider>
-            );
-
-        if (path.startsWith("/reserve/facility"))
-            return (
-                <FacilityListProvider>
-                    <FacilityListView />
-                </FacilityListProvider>
-            );
-
-        if (path.startsWith("/reserve/result")) {
-            return <ReserveResultView id={id} />;
-        }
-
-        if (path.startsWith("/reserve/review")) {
-            return (
-                <>
-                    <TitleBar name="리뷰 작성" />
-                    <ReviewInputFormView id={id} />
-                </>
-            );
-        }
-
-        if (path.startsWith("/reserve")) {
-            return id ? (
-                <ReserveDetailProvider>
-                    <TitleBar name="예약 상세" />
-                    <ReserveDetailView id={id} />
-                </ReserveDetailProvider>
-            ) : (
-                <ReserveListProvider>
-                    <ReserveListView />
-                </ReserveListProvider>
-            );
-        }
-
-        return (
-            <CenteredContainer>
-                <Typography>잘못된 경로입니다.</Typography>
-            </CenteredContainer>
-        );
-    };
+    // 에러 처리
 
     return (
-        <LocalizationProvider dateAdapter={AdapterDayjs}>
-            <ReserveRouteProvider>
-                <Container disableGutters sx={{ maxWidth: 500 }}>
-                    <RenderContent />
-                </Container>
-            </ReserveRouteProvider>
-        </LocalizationProvider>
+        <Container sx={{ maxWidth: 500 }}>
+            <Box
+                sx={{
+                    maxWidth: 500,
+                    width: "100%",
+                    display: "flex",
+                    justifyContent: "center",
+                    flexDirection: "column",
+                    gap: 2,
+                    position: "fixed",
+                    top: 122,
+                    left: "50%",
+                    transform: "translate(calc(-50%), -50%)", // 가운데 정렬 후 왼쪽으로 5px 더 이동
+                    background: "white",
+                    zIndex: 10,
+                }}
+            >
+                <CategoryFilter />
+                <Box my={2} sx={{ display: "flex", justifyContent: "row", pl: 2, pr: 2, gap: 31, mb: 0 }}>
+                    <SortFilter />
+                    <Button sx={{ bgcolor: "#FFF", borderRadius: 2 }} onClick={() => navigate("/reserve/list")}>
+                        내 예약 목록
+                    </Button>
+                </Box>
+                <Divider />
+            </Box>
+            <ListContent />
+        </Container>
     );
 };
+
+const Reserve = () => (
+    <LocalizationProvider dateAdapter={AdapterDayjs}>
+        <ReserveProvider>
+            <ReserveContent />
+        </ReserveProvider>
+    </LocalizationProvider>
+);
 
 export default Reserve;
