+++ src/services/reserveService.js
@@ -2,16 +2,22 @@ import instance from "./axiosInstance.js";
 
 const API_URL = "/reserve"; // 상대 URL
 
-export const getFacilityListToReserve = async (params = {}) => {
-    const { page = 0, size = 10, sortBy = "starPoint", category = "HOTEL", location } = params;
-    if (location.latitude === undefined || location.longitude === undefined) return;
-    return await instance
-        .get(`${API_URL}/facility/list`, {
+export const getFacilityListsToReserve = ({
+    page = 1,
+    size = 10,
+    sortBy = "starPoint",
+    category = "HOTEL",
+    location,
+    today,
+}) => {
+    return instance
+        .get(`${API_URL}/facility/lists`, {
             params: {
                 latitude: location.latitude,
                 longitude: location.longitude,
                 category,
                 sortBy,
+                day: today,
                 page,
                 size,
             },
@@ -23,9 +29,13 @@ export const getFacilityListToReserve = async (params = {}) => {
         });
 };
 
-export const getFacilityToReserveById = async ({ id }) => {
-    return await instance
-        .get(`${API_URL}/facility/detail/${id}`)
+export const getFacilityToReserveById = ({ id }) => {
+    return instance
+        .get(`${API_URL}/facility/detail`, {
+            params: {
+                id,
+            },
+        })
         .then((response) => response.data)
         .catch((error) => {
             console.error("시설 정보를 불러오는 데 실패했습니다.", error);
@@ -33,12 +43,11 @@ export const getFacilityToReserveById = async ({ id }) => {
         });
 };
 
-export const getFacilityReviewById = async ({ id, page, size }) => {
-    return await instance
-        .get(`${API_URL}/facility/${id}/review`, {
+export const getReserveListById = ({ id }) => {
+    return instance
+        .get(`${API_URL}/Reserve`, {
             params: {
-                page,
-                size,
+                id,
             },
         })
         .then((response) => response.data)
@@ -48,26 +57,12 @@ export const getFacilityReviewById = async ({ id, page, size }) => {
         });
 };
 
-export const getReserveList = async ({ page, size }) => {
-    return await instance
-        .get(`${API_URL}/list`, {
+export const getReserveDetailById = ({ uid, rid }) => {
+    return instance
+        .get(`${API_URL}/reserve/detail`, {
             params: {
-                page,
-                size,
-            },
-        })
-        .then((response) => response.data)
-        .catch((error) => {
-            console.error("예약 정보를 불러오는 데 실패했습니다.", error);
-            throw error; // 에러를 다시 던져서 호출한 곳에서 처리
-        });
-};
-
-export const getReserveDetailById = async ({ id }) => {
-    return await instance
-        .get(`${API_URL}/detail`, {
-            params: {
-                id,
+                uid,
+                rid,
             },
         })
         .then((response) => response.data)
